backend/python/requirements.txt
Flask==2.2.5
Flask-CORS==4.0.0
requests==2.31.0
azure-ai-textanalytics==5.3.0
azure-core==1.29.5
azure-cosmos==4.3.0
azure-ai-language-conversations==1.1.0
amadeus==8.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
